refactor: change imports, remove unnecessary code
