+++ src/components/UI/Tabs/Chat-tab.tsx
@@ -1,7 +1,7 @@
 import React, { ComponentProps, FC } from 'react';
 import { Tab } from './BaseTab/Tab';
 import { Message } from '../../../types/types';
-import { AvatarPlaceholder } from '../AvatarPlaceholder/AvatarPlaceholder';
+import { AvatarPlaceholder } from '../AvatarUI/AvatarPlaceholder';
 
 interface Props extends ComponentProps<'section'> {
   receiverName: string;
